\name{plot.d2}
\alias{plot.d2}
\title{Plotting sample correlation integrals}
\description{
Plotting method for objects inheriting from class '"d2"'.
}
\usage{
  \method{plot}{d2}(x, ...)
}
\arguments{
  \item{x}{'"d2"' object}
  \item{...}{additional graphical arguments}
}
\details{
Plots the sample correlation integrals in \code{x} in log-log scale, as a line for each considered embedding dimension.
}
\seealso{
\code{\link{d2}}
}
\author{Antonio, Fabio Di Narzo}
\keyword{hplot}
