\name{plot.false.nearest}
\alias{plot.false.nearest}
\title{Plotting false nearest neighbours results}
\description{
Plotting method for objects inheriting from class '"false.nearest"'.
}
\usage{
plot.false.nearest(x, ...)
}
\arguments{
  \item{x}{'"false.nearest"' object}
  \item{...}{additional graphical arguments}
}
\details{
Plots the results of \code{\link{false.nearest}}.
}
\seealso{
\code{\link{false.nearest}}
}
\author{Antonio, Fabio Di Narzo}
\keyword{hplot}
