\name{print.d2}
\alias{print.d2}
\title{Printing sample correlation integrals}
\description{
Printing method for objects inheriting from class '"d2"'.
}
\usage{
  \method{print}{d2}(x, ...)
}
\arguments{
  \item{x}{'"d2"' object}
  \item{...}{additional arguments to 'print'}
}
\details{
Simply calls \code{\link{plot.d2}}.
}
\seealso{
\code{\link{plot.d2}}, \code{\link{d2}}
}
\author{Antonio, Fabio Di Narzo}
\keyword{ts}
