\name{print.false.nearest}
\alias{print.false.nearest}
\title{Printing false nearest neighbours results}
\description{
Printing method for objects inheriting from class '"false.nearest"'.
}
\usage{
\method{print}{false.nearest}(x, ...)
}
\arguments{
  \item{x}{'"false.nearest"' object}
  \item{...}{additional arguments to 'print'}
}
\details{
Prints the table of results of \code{\link{false.nearest}}.
}
\seealso{
\code{\link{plot.false.nearest}}, \code{\link{false.nearest}}
}
\author{Antonio, Fabio Di Narzo}
\keyword{ts}
