\name{rossler.ts}
\alias{rossler.ts}
\title{Roessler simulated time series, without noise}
\description{
Roessler simulated time series, without noise. Of each state of the system, we observe the first component.
}
\details{
Roessler simulated time series, without noise, obtained with the call:

\code{rossler.ts <- sim.cont(rossler.syst, start=0, end=650, dt=0.1, start.x=c(0,0,0), parms=c(0.15, 0.2, 10))}
}
\author{Antonio, Fabio Di Narzo}
\keyword{datasets}
