\name{plot.ami}
\alias{plot.ami}
\title{Plotting average mutual information index}
\description{
Plotting method for objects inheriting from class '"ami"'.
}
\synopsis{
plot.ami(x, main=NULL, ...)
}
\usage{
plot.ami(x, ...)
}
\arguments{
  \item{x}{'"ami"' object}
  \item{...}{additional graphical arguments}
}
\details{
Plots the ami for each lag in \code{x}.
}
\seealso{
\code{\link{mutual}}
}
\author{Antonio, Fabio Di Narzo}
\keyword{hplot}
